backend/requirements.txt
# Flask và các extension
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
Pillow==10.0.1

# Xử lý ảnh và nhận diện khuôn mặt
face-recognition
opencv-python
dlib
scipy
numpy
imutils
scikit-learn

# Deep Learning (chọn 1 trong 2, tùy bạn dùng TensorFlow hay PyTorch)
tensorflow
# torch

# Tích hợp AI (OpenAI, Gemini)
openai
google-generativeai

# Báo cáo và xử lý dữ liệu
pandas
openpyxl

# Kết nối HTTP
requests

# PostgreSQL driver cho SQLAlchemy
psycopg2-binary
